self_discipline_lock_screen/requirements.txt
APScheduler==3.8.1
backports.zoneinfo==0.2.1
certifi==2020.6.20
cmake==3.22.1
comtypes==1.1.10
dlib==19.8.1
imageio==2.13.5
importlib-resources==5.4.0
imutils==0.5.4
numpy==1.19.5
opencv-python==4.5.5.62
Pillow==8.4.0
pyperclip==1.8.2
pypiwin32==223
PySimpleGUI==4.55.1
pyttsx3==2.90
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
six==1.16.0
tzdata==2021.5
tzlocal==4.1
wincertstore==0.2
zipp==3.6.0
